requirements.txt
amqp==2.5.2
Babel==2.8.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.10
django-celery-beat==1.6.0
django-celery-results==1.2.0
django-phonenumber-field==4.0.0
django-timezone-field==3.1
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
phonenumberslite==8.11.4
Pillow==7.0.0
PyMySQL==0.9.3
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
uWSGI==2.0.18
vine==1.3.0
zipp==2.2.0
